requirements.txt
crewai
crewai_tools
python-dotenv
google-generativeai
streamlit